requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.1.6
django-filter==21.1
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
Markdown==3.3.6
psycopg2-binary==2.9.3
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
